# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API for RockyDex Mobile App, supporting user authentication, favorites, and reading history synchronization.",
    "title": "RockyDex API",
    "contact": {},
    "version": "1.1.2"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/favorites": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns a list of all comics marked as favorite by the logged-in user.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Get user's favorite comics list",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/domain.Favorite"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Marks a comic as favorite.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Add a comic to user's favorites",
        "parameters": [
          {
            "description": "Favorite comic details",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.Favorite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Added to favorites",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/favorites/{slug}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Unmarks a comic as favorite by its slug.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Remove a comic from user's favorites",
        "parameters": [
          {
            "type": "string",
            "description": "Comic Slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Removed from favorites",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/favorites/check/{slug}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Checks if the user has favorited a comic.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Check if a comic is in user's favorites",
        "parameters": [
          {
            "type": "string",
            "description": "Comic Slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "is_favorite status",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "boolean"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/history": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns the chronological list of comics and chapters the user has read.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Get user's reading history",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/domain.History"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Saves or updates reading history and chapter progress for a comic.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Save reading history progress",
        "parameters": [
          {
            "description": "History details",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History saved",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Deletes all reading history records for the user.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Clear all reading history",
        "responses": {
          "200": {
            "description": "History cleared",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/history/{slug}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Deletes the user's reading history for a single comic.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Delete reading history for a specific comic",
        "parameters": [
          {
            "type": "string",
            "description": "Comic Slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "History deleted",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/history/comic/{slug}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns the reading history for a specific comic if it exists.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "library"
        ],
        "summary": "Get reading history for a specific comic",
        "parameters": [
          {
            "type": "string",
            "description": "Comic Slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/domain.History"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "domain.Favorite": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "string"
        },
        "comic_slug": {
          "type": "string"
        },
        "comic_name": {
          "type": "string"
        },
        "comic_thumb": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        }
      }
    },
    "domain.History": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "string"
        },
        "comic_slug": {
          "type": "string"
        },
        "comic_name": {
          "type": "string"
        },
        "comic_thumb": {
          "type": "string"
        },
        "chapter_slug": {
          "type": "string"
        },
        "chapter_name": {
          "type": "string"
        },
        "progress_percent": {
          "type": "integer"
        },
        "last_read_at": {
          "type": "string"
        }
      }
    }
  }
}